Move LCSDiff to Differ companion object
